Make native build compile against current upstream
Upstream main rebased a few things that broke the native build (every
native CI run since the MCGA addition has failed):

- scanout_context.h: base vramSize() declared as uint32_t, but all 5
  subclass overrides (cga/ega/hercules/tandy/mcga) return size_t. GCC
  rejects mismatched primitive return types in overrides. Aligning the
  base to size_t is the smallest fix and matches the subclasses.
- platformio.ini: add video/mcga.cpp to the native build_src_filter so
  the new MCGA implementation links.
- native/native_driver_stubs.cpp: VideoScanout's m_lineOffset field was
  removed and m_vramSize added upstream; updated the stub initializer
  list to match. Added a stub for the new drawScanline_mcga_320x200x256
  static method.

+++ src/video/scanout_context.h
@@ -23,6 +23,7 @@
 
 #include "drivers/video/display.h"
 
+#include <stddef.h>
 #include <stdint.h>
 
 using fabgl::RGB222;
@@ -44,7 +45,7 @@ public:
   virtual uint8_t *plane(uint8_t index) { return nullptr; }
 
   // Total VRAM size in bytes, used for address wrapping
-  virtual uint32_t vramSize() = 0;
+  virtual size_t vramSize() = 0;
 
   // CRTC start address, expressed in words
   virtual uint32_t startAddress() = 0;
